feat: ImageBox disabled 속성 추가
